# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photonics Engine Simulator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ComponentType = (typeof ComponentType)[keyof typeof ComponentType];

export const ComponentType = {
  laser_source: "laser_source",
  waveguide: "waveguide",
  beam_splitter: "beam_splitter",
  coupler: "coupler",
  modulator: "modulator",
  photodetector: "photodetector",
  optical_amplifier: "optical_amplifier",
  phase_shifter: "phase_shifter",
  filter: "filter",
  isolator: "isolator",
  circulator: "circulator",
  mzi: "mzi",
  ring_resonator: "ring_resonator",
  grating_coupler: "grating_coupler",
  mirror: "mirror",
} as const;

export type ComponentTemplateParameterSchema = { [key: string]: unknown };

export interface ComponentParams {
  /** Operating wavelength in nm */
  wavelength?: number;
  /** Optical power in dBm */
  power?: number;
  /** Insertion loss in dB */
  loss?: number;
  /** Power split ratio (0-1) */
  splitRatio?: number;
  /** Coupling coefficient (0-1) */
  couplingCoeff?: number;
  /** Component length in um */
  length?: number;
  /** Effective refractive index */
  neff?: number;
  /** Propagation loss in dB/cm */
  alpha?: number;
  /** Gain in dB */
  gain?: number;
  /** Detector responsivity in A/W */
  responsivity?: number;
  /** Phase shift in radians */
  phaseShift?: number;
  /** Bandwidth in GHz */
  bandwidth?: number;
  /** Modulation extinction ratio in dB */
  extinctionRatio?: number;
  /** Mirror reflectivity (0-1) */
  reflectivity?: number;
}

export interface ComponentTemplate {
  type: ComponentType;
  label: string;
  description: string;
  defaultParams: ComponentParams;
  parameterSchema?: ComponentTemplateParameterSchema;
}

export interface CircuitComponent {
  id: string;
  type: ComponentType;
  label: string;
  x: number;
  y: number;
  params: ComponentParams;
}

export interface Connection {
  id: string;
  fromComponentId: string;
  fromPort: string;
  toComponentId: string;
  toPort: string;
}

export interface CircuitLayout {
  components: CircuitComponent[];
  connections: Connection[];
}

export type BuildStatus = (typeof BuildStatus)[keyof typeof BuildStatus];

export const BuildStatus = {
  draft: "draft",
  simulating: "simulating",
  converged: "converged",
  needs_revision: "needs_revision",
} as const;

export interface Build {
  id: number;
  name: string;
  description?: string;
  /** Target operating wavelength in nm */
  targetWavelength: number;
  /** Target output power in dBm */
  targetPower?: number;
  /** Target signal-to-noise ratio in dB */
  targetSNR?: number;
  layout: CircuitLayout;
  iterationCount: number;
  status: BuildStatus;
  /** Score 0-100 indicating how close to equilibrium */
  equilibriumScore: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBuildRequest {
  name: string;
  description?: string;
  targetWavelength: number;
  targetPower?: number;
  targetSNR?: number;
  layout: CircuitLayout;
}

export interface UpdateBuildRequest {
  name?: string;
  description?: string;
  targetWavelength?: number;
  targetPower?: number;
  targetSNR?: number;
  layout?: CircuitLayout;
}

export type ComponentResultStatus =
  (typeof ComponentResultStatus)[keyof typeof ComponentResultStatus];

export const ComponentResultStatus = {
  ok: "ok",
  warning: "warning",
  error: "error",
} as const;

export type IssueSeverity = (typeof IssueSeverity)[keyof typeof IssueSeverity];

export const IssueSeverity = {
  info: "info",
  warning: "warning",
  error: "error",
} as const;

export interface Issue {
  code: string;
  severity: IssueSeverity;
  message: string;
  suggestion: string;
  componentId?: string;
}

export interface ComponentResult {
  componentId: string;
  label: string;
  type: ComponentType;
  inputPower: number;
  outputPower: number;
  phase: number;
  wavelength: number;
  loss: number;
  gain?: number;
  status: ComponentResultStatus;
  issues: Issue[];
}

export interface SimulationResult {
  id: number;
  buildId: number;
  iterationNumber: number;
  totalInputPower: number;
  totalOutputPower: number;
  systemLoss: number;
  snr: number;
  coherenceLength: number;
  wavelength: number;
  /** Score 0-100 */
  equilibriumScore: number;
  componentResults: ComponentResult[];
  issues: Issue[];
  converged: boolean;
  suggestions: string[];
  createdAt: string;
}