# lib/api-zod/src/generated/types/build.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photonics Engine Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { BuildStatus } from "./buildStatus";
import type { CircuitLayout } from "./circuitLayout";

export interface Build {
  id: number;
  name: string;
  description?: string;
  /** Target operating wavelength in nm */
  targetWavelength: number;
  /** Target output power in dBm */
  targetPower?: number;
  /** Target signal-to-noise ratio in dB */
  targetSNR?: number;
  layout: CircuitLayout;
  iterationCount: number;
  status: BuildStatus;
  /** Score 0-100 indicating how close to equilibrium */
  equilibriumScore: number;
  createdAt: Date;
  updatedAt: Date;
}