# lib/api-zod/src/generated/types/circuitComponent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photonics Engine Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { ComponentParams } from "./componentParams";
import type { ComponentType } from "./componentType";

export interface CircuitComponent {
  id: string;
  type: ComponentType;
  label: string;
  x: number;
  y: number;
  params: ComponentParams;
}