# lib/api-zod/src/generated/types/circuitLayout.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photonics Engine Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { CircuitComponent } from "./circuitComponent";
import type { Connection } from "./connection";

export interface CircuitLayout {
  components: CircuitComponent[];
  connections: Connection[];
}