# lib/api-zod/src/generated/types/componentParams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photonics Engine Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface ComponentParams {
  /** Operating wavelength in nm */
  wavelength?: number;
  /** Optical power in dBm */
  power?: number;
  /** Insertion loss in dB */
  loss?: number;
  /** Power split ratio (0-1) */
  splitRatio?: number;
  /** Coupling coefficient (0-1) */
  couplingCoeff?: number;
  /** Component length in um */
  length?: number;
  /** Effective refractive index */
  neff?: number;
  /** Propagation loss in dB/cm */
  alpha?: number;
  /** Gain in dB */
  gain?: number;
  /** Detector responsivity in A/W */
  responsivity?: number;
  /** Phase shift in radians */
  phaseShift?: number;
  /** Bandwidth in GHz */
  bandwidth?: number;
  /** Modulation extinction ratio in dB */
  extinctionRatio?: number;
  /** Mirror reflectivity (0-1) */
  reflectivity?: number;
}