# lib/api-zod/src/generated/types/componentResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photonics Engine Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { ComponentResultStatus } from "./componentResultStatus";
import type { ComponentType } from "./componentType";
import type { Issue } from "./issue";

export interface ComponentResult {
  componentId: string;
  label: string;
  type: ComponentType;
  inputPower: number;
  outputPower: number;
  phase: number;
  wavelength: number;
  loss: number;
  gain?: number;
  status: ComponentResultStatus;
  issues: Issue[];
}