# lib/api-zod/src/generated/types/componentType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photonics Engine Simulator API
 * OpenAPI spec version: 0.1.0
 */

export type ComponentType = (typeof ComponentType)[keyof typeof ComponentType];

export const ComponentType = {
  laser_source: "laser_source",
  waveguide: "waveguide",
  beam_splitter: "beam_splitter",
  coupler: "coupler",
  modulator: "modulator",
  photodetector: "photodetector",
  optical_amplifier: "optical_amplifier",
  phase_shifter: "phase_shifter",
  filter: "filter",
  isolator: "isolator",
  circulator: "circulator",
  mzi: "mzi",
  ring_resonator: "ring_resonator",
  grating_coupler: "grating_coupler",
  mirror: "mirror",
} as const;