# lib/api-zod/src/generated/types/connection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photonics Engine Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface Connection {
  id: string;
  fromComponentId: string;
  fromPort: string;
  toComponentId: string;
  toPort: string;
}