# lib/api-zod/src/generated/types/createBuildRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photonics Engine Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { CircuitLayout } from "./circuitLayout";

export interface CreateBuildRequest {
  name: string;
  description?: string;
  targetWavelength: number;
  targetPower?: number;
  targetSNR?: number;
  layout: CircuitLayout;
}