# lib/api-zod/src/generated/types/issue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photonics Engine Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { IssueSeverity } from "./issueSeverity";

export interface Issue {
  code: string;
  severity: IssueSeverity;
  message: string;
  suggestion: string;
  componentId?: string;
}