# lib/api-zod/src/generated/types/simulationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photonics Engine Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { ComponentResult } from "./componentResult";
import type { Issue } from "./issue";

export interface SimulationResult {
  id: number;
  buildId: number;
  iterationNumber: number;
  totalInputPower: number;
  totalOutputPower: number;
  systemLoss: number;
  snr: number;
  coherenceLength: number;
  wavelength: number;
  /** Score 0-100 */
  equilibriumScore: number;
  componentResults: ComponentResult[];
  issues: Issue[];
  converged: boolean;
  suggestions: string[];
  createdAt: Date;
}